处理一些问题

+++ src/main/java/com/keymanager/ckadmin/dao/QZOperationTypeDao.java
@@ -28,7 +28,7 @@ public interface QZOperationTypeDao extends BaseMapper<QZOperationType> {
     /**
      * 上一级的主键
      */
-    int selectLastId();
+    long selectLastId();
 
     void updateQZOperationTypeStandardTime(@Param("uuid") long uuid, @Param("isStandardFlag") int isStandardFlag);
